==> project-3/code/part2_cleaning.ipynb <==
# Project 3: Advanced Running Retargeting using NLP

---


## Part 2: Initial Cleaning

This section involves initial cleaning of the raw Reddit datasets. The original raw r/AdvancedRunning dataset contained 13454 rows and 92 columns, while the original raw r/C25K dataset contained 15037 rows and 96 columns. 'Removed' posts were dropped from both datasets. New data frames were created for both subreddits containing 'author', 'title' and 'selftext' (post). The 'title' and 'selftext' columns were merged into one 'post' column. Rows with missing values were then identified and dropped from both datasets. Duplicate posts were then dropped from both datasets. Reddit moderators were identified and moderator posts were dropped from each dataset. After initial cleaning, the r/AdvancedRunning dataset contained 8440 rows and the r/C25K contained 5132 rows. A target column for classification was added to each dataset, where '1' indicates r/AdvancedRunning and '0' indicates r/C25k. The initially cleaned datasets were then concatenated into one dataset to be used for pre-processing in the following section.

#imports
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

#Read in raw csv

raw_advanced_running=pd.read_csv('../data/raw_advanced_running.csv', low_memory=False)
raw_couch_5k=pd.read_csv('../data/raw_couch_5k.csv', low_memory=False)

#### Initial Data Inspection

#Look at data
raw_advanced_running.head()

#Look at data
raw_couch_5k.head()

#View shape of data
print(raw_advanced_running.shape)
print(raw_couch_5k.shape)

#### Removing all posts that contained 'removed'

#Checking and removing [removed] and re-assining to df without removed
raw_advanced_running=raw_advanced_running.loc[raw_advanced_running['selftext']!='[removed]']
raw_couch_5k=raw_couch_5k.loc[raw_couch_5k['selftext']!='[removed]']

#### Creating new DataFrames with wanted columns
1. Author, Title and Post
2. Making Title and Post into one all-text column 

#Making new dataframes
raw_advanced=pd.DataFrame()
raw_couch5k=pd.DataFrame()

#Creating dataframes with wanted information 
raw_advanced['author']=raw_advanced_running['author']
raw_advanced['post']=raw_advanced_running['title'] + ' ' + raw_advanced_running['selftext']
raw_couch5k['author']=raw_couch_5k['author']
raw_couch5k['post']=raw_couch_5k['title'] + ' ' + raw_couch_5k['selftext']

#Checking 
raw_advanced.head()

#Checking
raw_couch5k.head()

#### Checking for Null Values and Data Types

#Checking for nulls and datatypes 
raw_advanced.info()

#Checking for nulls and datatypes 
raw_couch5k.info()

#### Dropping Missing Values 

#Removing rows with missing values 
raw_advanced.dropna(inplace=True)
raw_couch5k.dropna(inplace=True)

#Checking
raw_advanced.info()

#Checking
raw_couch5k.info()

#### Dropping duplicate posts in each dataset. Indicating to keep the first post of every duplicate. 

#Dropping duplicates 
raw_advanced.drop_duplicates(subset='post', keep='first', inplace=True)

raw_couch5k.drop_duplicates(subset='post', keep='first', inplace=True)

#Checking shape after dropping duplicates
print(raw_advanced.shape)
print(raw_couch5k.shape)

#### Dropping moderator posts from each dataset
- Moderator's were identified in each subreddit and rows containing moderator posts were dropped from both datasets

#Dropping mods by selecting rows with mods in both df
raw_advanced.drop(raw_advanced.index[(raw_advanced['author'] == 'brwalkernc') | 
                             (raw_advanced['author'] == 'CatzerzMcGee') | 
                            (raw_advanced['author'] == 'aewillia') | 
                             (raw_advanced['author'] == 'AutoModerator') |
                            (raw_advanced['author'] == 'ruinawish')], inplace=True)

raw_couch5k.drop(raw_couch5k.index[(raw_couch5k['author'] == 'AshKals') | 
                           (raw_couch5k['author'] == 'cainunable') | 
                            (raw_couch5k['author'] == 'C25k_bot')], inplace=True)

#### Checking final shape for each dataframe
- Both subreddit's were within an appropriate amount for modeling. Extra care will be taken during modeling to deal with imbalanced data

#Checking final shape of both dfs
print(raw_advanced.shape)
print(raw_couch5k.shape)

#### Binary column created for target classification column 

#Creating 1 and 0 columns for each dataset used for classification 
raw_advanced['is_advanced']=1
raw_couch5k['is_advanced']=0

#### Both datasets were then concatenated together into one dataset to be used for pre-processing

#concat subreddits together 
clean_runners=pd.concat([raw_advanced, raw_couch5k], axis=0)
clean_runners.tail()

#Saving cleaned dataset
clean_runners.to_csv('clean_runners.csv', index=False)

---
#### Next Section: Part 3 - Pre-Processing
